// frontend/src/app/domain/models/datos-entrada.interface.ts
export interface DatosEntrada {
  price: number;
  freight_value: number;
  product_name_lenght: number;
  product_description_lenght: number;
  product_photos_qty: number;
  product_weight_g: number;
  product_length_cm: number;
  product_height_cm: number;
  product_width_cm: number;
  payment_sequential: number;
  payment_installments: number;
  payment_value: number;
}